Add check for gaps in roll tables
It is possible that a table CSV may miss out a number, meaning that if
that number is rolled the program would not know which event to return.
Such gaps are not usual for roll tables and uninteresting ranges of
rolls should be explicitly labelled as such e.g. "30-60,Nothing
happens".

table_loaded now checks to see if there are gaps in the table loaded
from file and raises an error if there are.