Fix check for gaps in table in load_table
itertools.chain takes each iterable as a separate argument, so
table.keys() must be unpacked rather than being passed to chain as a
single argument.

The chain iterator is cast to a list to prevent depletion so that both
the minimum and maximum values can be extracted. The list is sorted
because the keys of a dict are not returned in any order and so
rolls_in_table might be in a different order than the range, meaning
that they would be unequal even if they contained the same elements.