Add GUI for loading and rolling against tables
Added gui.py which contains a class implementing a GUI for the roller.

When loading files, it is assumed that they will not have a header row.
If no satisfactory interface for specifying whether files being loaded
have a header row or not, then the headers option in
table_handling.load_table should be removed and the presence or lack of
headers should be incorporated into the required format.

The GUI currently has no way to handle errors, such as IOErrors when
opening files, gracefully.