Remove headers option from load_table
As mentioned in 5592556, the GUI for the table roller has no way of
letting the user specify whether a file's first line is a table entry or
merely a header line. No satisfactory way to allow the user to specify
this can be found, so the headers option has been removed from
load_table and headers are now a mandatory part of the format for CSV
table files.