Remove headers kwarg from load_table
As of 5592556, the GUI assumed that the top row of table files would
not contain headers but actual table entries. No satisfactory way to
allow the user to specify whether a file contains headers can be found.
Therefore, the headers option has been removed from load_table and it is
now assumed/required that the first row of a table file contains
headers.