Require that header line starts with '%'
It is necessary that we know that the first row of a table file
is a header row and not a table entry. It may be possible to regex the
line to see if it a valid entry, but it is much easier to simply
require that the header row starts with a marker of some sort. For
this, '%' is chosen because it cannot be part of a valid roll/roll
range.

By requiring that the header file start for '%', we can ensure that the
user has begun the file with headers.