Require that roll ranges do not overlap
For a long time, it has been stated in load_table's docstring that
roll ranges should not overlap i.e. there should not be two events
associated with one roll, but this was not checked for. load_table
now checks that none of the ranges overlap and raises
TableFormatError if they do.