Add check for repeated roll ranges
Previously, it was permitted to repeat a roll range as long as it did
not make a gap in the table. Now load_table checks to see if a range is
already in the table before adding it and raises TableFormatError if it
is.

This change makes possible a fix to the overlap checking in load_table.
itertools.product returns pairs that are two of the same range, meaning
they overlap but in reality it doesn't matter. Explicitly checking for
duplicates means that we can safely ignore these 'twins' when checking
for overlaps.