Fix overlap checking on identical pairs
itertools.product returns pairs that are two of the same range, meaning
they overlap. load_table therefore raised an exception even though the
two ranges are the exact same one and their overlapping isn't a problem.
This is fixed by ignoring pairs of equal ranges, which we can safely do
because of the explicit check for duplicates introduced in commit a6f3676.